requirements.txt
google_api_python_client==2.102.0
google_auth_oauthlib==1.1.0
numpy==1.25.2
pandas==2.0.3
protobuf==4.25.1
python-dotenv==0.19.0
selenium==4.16.0
streamlit==1.29.0
